Refactor: Вынес логику точек маршрута в отдельный PointPresenter, реализовал поддержку редактирования и избранного
