+++ package.json
@@ -13,9 +13,11 @@
     "@babel/preset-env": "^7.26.9",
     "babel-loader": "^10.0.0",
     "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^7.1.2",
     "eslint": "8.38.0",
     "eslint-config-htmlacademy": "9.0.0",
     "html-webpack-plugin": "^5.6.3",
+    "style-loader": "^4.0.0",
     "webpack": "^5.99.0",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.2.1"
