+++ src/const.js
@@ -1,8 +1,8 @@
 export const POINT_TYPES = [
-    'taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'
-  ];
+  'taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'
+];
 
-  export const SortType = {
+export const SortType = {
   DAY: 'day',
   PRICE: 'price',
   TIME: 'time',
