+++ src/main.js
@@ -1,6 +1,6 @@
 
 import TripPresenter from './presenter/trip-presenter.js';
- 
+
 document.addEventListener('DOMContentLoaded', () => {
   const tripPresenter = new TripPresenter();
   tripPresenter.init();
