+++ src/model/destination.js
@@ -1,8 +1,8 @@
 export default class Destination {
-    constructor(id, name, description, pictures) {
-      this.id = id;
-      this.name = name;
-      this.description = description;
-      this.pictures = pictures;
-    }
+  constructor(id, name, description, pictures) {
+    this.id = id;
+    this.name = name;
+    this.description = description;
+    this.pictures = pictures;
   }
+}
