+++ src/model/offer.js
@@ -1,7 +1,7 @@
 export default class Offer {
-    constructor(id, title, price) {
-      this.id = id;
-      this.title = title;
-      this.price = price;
-    }
+  constructor(id, title, price) {
+    this.id = id;
+    this.title = title;
+    this.price = price;
   }
+}
