+++ src/presenter/PointPresenter.js
